Edit Task in Project and Display Same Project
Created function EditTaskInProject() which requires reference to
current project (globally) as well as current task to be edited
(globally). The find() method matches the task in the project task
list with the current task. This matching task is then updated with
new values. The entire project tasks list is rerendered.