Set Color of Task Based on Priority
Obtain reference to added task/edited task's priority value, and assign
to variable 'priorityOfCurrentTask' in changingHTMLElement.js. Obtain
reference to added task/edited task, and search for the matching HTML
Element (match based on title). Use switch block to add a class to the
matching HTML element, with the class determining the background color
of the task.